usm, npm: classification: rename file
